refactor: 역직렬화 기능 분리
