--- tomcat/src/main/java/org/apache/Serializer.java
@@ -1,10 +1,0 @@
-package org.apache;
-
-import org.apache.coyote.HttpResponse;
-
-public class Serializer {
-
-    public static byte[] serialize(HttpResponse response) {
-        return response.getResponseMessage().getBytes();
-    }
-}
